src/models/text_engine.py
import torch
from transformers import LogitsProcessor

class DNAWatermarkLogitsProcessor(LogitsProcessor):
    """
    LogitsProcessor for LLM Text DNA-Stamping.
    Uses the "Soft Watermarking" technique (Kirchenbauer et al.).
    Biases token selection towards a 'green list' derived from DNA bits.
    """
    def __init__(self, dna_bits: torch.Tensor, gamma: float = 0.5, delta: float = 2.0):
        """
        Args:
            dna_bits: The 128-bit DNA vector from Mojo Core.
            gamma: Fraction of vocabulary in the green list.
            delta: The 'bias' added to green tokens.
        """
        self.dna_bits = dna_bits
        self.gamma = gamma
        self.delta = delta
        self.rng = torch.Generator()

    def _get_green_list(self, prev_token_id: int, vocab_size: int) -> torch.Tensor:
        # Use previous token + DNA bits to seed the green list selection
        # This makes the watermark dependent on the DNA bits
        seed = prev_token_id + int(self.dna_bits.sum().item() * 1000)
        self.rng.manual_seed(seed)
        
        # Create a split of the vocab
        perm = torch.randperm(vocab_size, generator=self.rng)
        green_list_size = int(vocab_size * self.gamma)
        return perm[:green_list_size]

    def __call__(self, input_ids: torch.LongTensor, scores: torch.FloatTensor) -> torch.FloatTensor:
        # Get vocab size from scores
        vocab_size = scores.shape[-1]
        
        for i in range(input_ids.shape[0]):
            prev_token = input_ids[i, -1].item()
            green_indices = self._get_green_list(prev_token, vocab_size)
            
            # Apply the bias (The 'Stamp')
            scores[i, green_indices] += self.delta
            
        return scores

class TextDNAEngine:
    """
    High-level engine for Text DNA-Stamping.
    Connects the Mojo DNA bits to LLM generation.
    """
    def __init__(self):
        # We'd typically load a tokenizer/model here
        pass

    def get_logits_processor(self, dna_bits: torch.Tensor):
        return DNAWatermarkLogitsProcessor(dna_bits)

    def extract_watermark(self, text: str, dna_bits: torch.Tensor) -> float:
        """
        Statistical verification of the DNA signature in a text sample.
        Returns a 'z-score' of confidence.
        """
        # (This would implement the counting logic for green tokens)
        return 0.99 
