pyproject.toml
[project]
name = "stegadna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "mojo>=0.26.1.0",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.2",
    "uvicorn>=0.40.0",
    "wandb>=0.24.1",
]
